+++ packages/iov-bns/types/tags.d.ts
@@ -1,2 +1,2 @@
-import { AtomicSwapQuery, BcpQueryTag } from "@iov/bcp";
-export declare function bnsSwapQueryTag(query: AtomicSwapQuery, set?: boolean): BcpQueryTag;
+import { AtomicSwapQuery, QueryTag } from "@iov/bcp";
+export declare function bnsSwapQueryTag(query: AtomicSwapQuery, set?: boolean): QueryTag;