+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Breaking changes
 - @iov/bcp: Change type of `Account.balance` to array of `Amount`, i.e. remove
   the `tokenName` field.
 - @iov/bcp: Remove obsolete type `BcpCoin`.
+- @iov/bcp: Rename `BcpQueryTag` to `QueryTag` and `BcpTxQuery` to
+  `TransactionQuery`.
 
 ## 0.13.3